Change UI of add course
